fix router issues and reorganize the code
* We split the routing code into two modules. One contains the
  concrete route definitions and loader function. The other contains the
  routing logic that is agnostic of the concrete router types.
* We don’t persist `LoadedRoute`. Instead we persist `Route` and load on
  every push or pop. As a consequence, `activeRouterStore` is not a
  derived store anymore.
* Consolidate functionality in `setHistory`
* Extract `unreachable`. It’s not specific to the router
