+++ ui/src/unreachable.ts
@@ -1,0 +1,9 @@
+import * as error from "ui/src/error";
+
+export const unreachable = (value: never): never => {
+  throw new error.Error({
+    code: error.Code.Unreachable,
+    message: "Unreachable code",
+    details: { value },
+  });
+};
